refactor: move `prev` command to separate file
